+++ Production/govuk_ios/Extensions/UserNotifications/UNUserNotificationCenter+Interface.swift
@@ -1,0 +1,14 @@
+import Foundation
+import UserNotifications
+
+protocol UserNotificationCenterInterface {
+    var authorizationStatus: UNAuthorizationStatus { get async }
+}
+
+extension UNUserNotificationCenter: UserNotificationCenterInterface {
+    var authorizationStatus: UNAuthorizationStatus {
+        get async {
+            await notificationSettings().authorizationStatus
+        }
+    }
+}
