+++ Production/govuk_ios/Services/NotificationService.swift
@@ -60,15 +60,3 @@ class NotificationService: NotificationServiceInterface {
         completion(.success(slides))
     }
 }
-
-protocol UserNotificationCenterInterface {
-    var authorizationStatus: UNAuthorizationStatus { get async }
-}
-
-extension UNUserNotificationCenter: UserNotificationCenterInterface {
-    var authorizationStatus: UNAuthorizationStatus {
-        get async {
-            await notificationSettings().authorizationStatus
-        }
-    }
-}
